+++ CHANGES.rst
@@ -5,6 +5,9 @@ Changelog
 2.0.0rc6 (unreleased)
 ---------------------
 
+- add booking_code field to IPrenotazione schema
+  update locales
+  [lucabel]
 - Traduzioni
   [mamico]